blockifier_reexecution: move SimulatedStateReader and serde utils from starknet_transaction_prover (#13654)
